@@ -12,16 +12,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.2.5",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "7.2.5",
-    "@angular/compiler": "7.2.5",
-    "@angular/core": "7.2.5",
-    "@angular/forms": "7.2.5",
-    "@angular/http": "7.2.5",
-    "@angular/platform-browser": "7.2.5",
-    "@angular/platform-browser-dynamic": "7.2.5",
-    "@angular/router": "7.2.5",
+    "@angular/animations": "8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "8.2.14",
+    "@angular/forms": "8.2.14",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/router": "8.2.14",
     "@auth0/angular-jwt": "^2.0.0",
     "chart.js": "^2.9.3",
     "core-js": "^2.4.1",
@@ -31,15 +30,15 @@
     "ng2-toasty": "^4.0.3",
     "primeicons": "^2.0.0",
     "primeng": "7.1.3",
-    "rxjs": "^6.2.1",
+    "rxjs": "^6.5.5",
     "tslib": "^1.9.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "^7.3.10",
-    "@angular/compiler-cli": "7.2.5",
-    "@angular/language-service": "7.2.5",
+    "@angular-devkit/build-angular": "~0.803.26",
+    "@angular/cli": "^8.3.26",
+    "@angular/compiler-cli": "8.2.14",
+    "@angular/language-service": "8.2.14",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
@@ -54,7 +53,7 @@
     "protractor": "^5.4.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "3.2.4"
+    "typescript": "3.5.3"
   },
   "engines": {
     "node": "12.16.2",
